+++ js/plugins/dropdown.js
@@ -9,8 +9,6 @@ var Dropdown = {
         this._setOptionsFromDOM();
         this._create();
 
-        Utils.exec(this.options.onDropdownCreate, [this.element]);
-
         return this;
     },
 
@@ -39,10 +37,20 @@ var Dropdown = {
     },
 
     _create: function(){
-        var that = this, element = this.element, o = this.options;
-        var toggle, parent = element.parent();
-        var element_roles = Utils.isValue(element.attr("data-role")) ? Utils.strToArray(element.attr("data-role")) : [];
+        var that = this;
+        this._createStructure();
+        this._createEvents();
+        Utils.exec(this.options.onDropdownCreate, [this.element]);
+        if (this.element.hasClass("open")) {
+            setTimeout(function(){
+                that.open();
+            }, 500)
+        }
+    },
 
+    _createStructure: function(){
+        var that = this, element = this.element, o = this.options;
+        var toggle;
         toggle = o.toggleElement !== null ? $(o.toggleElement) : element.siblings('.dropdown-toggle').length > 0 ? element.siblings('.dropdown-toggle') : element.prev();
 
         this.displayOrigin = element.css("display");
@@ -53,10 +61,12 @@ var Dropdown = {
 
         element.css("display", "none");
 
-        if (element_roles.length === 0 || element_roles.indexOf("dropdown") === -1) {
-            element_roles.push("dropdown");
-            element.attr("data-role", element_roles.join(", "));
-        }
+        this._toggle = toggle;
+    },
+
+    _createEvents: function(){
+        var that = this, element = this.element, o = this.options;
+        var toggle = this._toggle, parent = element.parent();
 
         toggle.on(Metro.events.click, function(e){
             parent.siblings(parent[0].tagName).removeClass("active-container");
@@ -92,8 +102,6 @@ var Dropdown = {
             e.preventDefault();
             e.stopPropagation();
         });
-
-        this._toggle = toggle;
 
         if (o.noClose === true) {
             element.addClass("keep-open").on(Metro.events.click, function (e) {